refactor(examples): move coding cli capabilities
